requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
supabase==1.2.0
scikit-learn==1.3.2
python-dotenv==1.0.0
lxml==4.9.3
pandas==2.1.4
numpy==1.24.3
pytest==7.4.3
flask==2.3.3
flask-cors==4.0.0
networkx==3.2

